+++ test/integer_neg.cpp
@@ -39,7 +39,7 @@ struct neg_tester {
         mpz_raii m1, m2;
         integer n1, n2;
         ::mpz_neg(&m1.m_mpz, &m2.m_mpz);
-        neg(n1, n2);
+        REQUIRE(&neg(n1, n2) == &n1);
         REQUIRE((lex_cast(n1) == lex_cast(m1)));
         REQUIRE(n1.is_static());
         // Test the other variants.